sprint 0 - SPRINT 0 — CONSOLIDAÇÃO OBRIGATÓRIA DO DOMÍNIO

+++ internal/domain/payment/payment_summary.go
@@ -1,9 +1,9 @@
 package payment
 
 type PaymentSummary struct {
-	TotalPaid    float64 `json:"total_paid"`
-	TotalPending float64 `json:"total_pending"`
-	TotalExpired float64 `json:"total_expired"`
+	TotalPaid    int64 `json:"total_paid"`
+	TotalPending int64 `json:"total_pending"`
+	TotalExpired int64 `json:"total_expired"`
 
 	CountPaid    int64 `json:"count_paid"`
 	CountPending int64 `json:"count_pending"`
